requirements.txt
python-telegram-bot>=20.0
gspread>=5.0
google-auth>=2.0
requests>=2.28
python-dotenv>=1.0


